+++ CHANGELOG.md
@@ -14,6 +14,8 @@ way to update this template, but currently, we follow a pattern:
 
 ## Upcoming version 2019-XX-XX
 
+- [fix] Menu needs to wait for mounting to calculate dimensions properly.
+  [#1096](https://github.com/sharetribe/flex-template-web/pull/1096)
 - [fix] Renamed Component.example.css files to ComponentExample.css to fix bug introduced in one of
   the library updates. [#1095](https://github.com/sharetribe/flex-template-web/pull/1095)
 - [add] `rawOnly` flag for Styleguide examples using fixed positioning or full-page dimensions.